requirements.txt
numpy
tqdm
transformers
torch
networkx
openai
python-dotenv
requests 